pyproject.toml
[project]
name = "gap-robots-frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pyqt5>=5.15.11",
    "requests>=2.32.4",
]
